fix: wrong buffer write callback method
Since cURL will call this function for many times if the data is too long (>1K?), directly
write things to buffer will cause content overlapping.

Use a pointer to point at the end of buffer (if something in it), and write data towards the
pointer `tmp` which points to the end of buffer.

Signed-off-by: lmy441900 <lmy441900@gmail.com>